DateTimePickerFragment - Initialize clean up on cancel or dismiss.
The date picker only initialize the clean up if it is cancelled
because we still need the DateTimePickerFragment when a date have
been selected, i.e. the date picker is dismissed.

The time picker however needs to initialize the clean up when
it's dismissed, since it doesn't matter why it was dismissed.