Redesigned the interaction between activity and fragment when creating new project.
I've redesigned the interaction between the MainActivity and
ProjectsFragment when creating new projects. This is done by
dispatching the create event to the fragment and let it deal
with the dialog and adapter update.

Also, it makes the connection between the activity and fragment
a bit more loosely coupled since the dependencies have been
reduced and only relys on an interface.