Changed subscribe/unsubscribe for the refresh of active projects.
Because the detachView-method in the MvpPresenter is called when
the view is actually destroyed the subscription for refreshing
active projects is still active when the user navigates to another
activity (within the application).

This is really unnecessary as the fragment is not displayed.
A better solution is to have it subscribe in the onResume-method,
and unsubscribe in the onPause-method.