Update unit test for Project domain model
I've updated the unit tests for the Project domain model. Most of the
changes is not to the actual unit test but the structure around them.

The major points of the structural changes are as follows.

Renamed the test methods. Previously, I was inconsistent about naming
the methods, so I've tried to some degree to adapt the names to a new
naming convention, "MethodName_StateUnderTest_ExpectedBehaviour".

Rearranged methods in order of expected behaviour, you have to fail
before you can succeed.

Changed all primitive data types to their respective object wrappers.
The reason for this is, the domain models should work with objects and
not primitives where possible.

I've also added some additional tests for cases I've missed before and
tweaked some tests to be more consistent.