Update unit test for Time domain model
I've updated the unit tests for the Time domain model. Most of the
changes are not to the actual unit test but the structure around them.

The major points of the structural changes are as follows.

Renamed the test methods. Previously, I was inconsistent about naming
the methods, so I've tried to some degree to adapt the names to a new
naming convention, "MethodName_StateUnderTest_ExpectedBehaviour".

Rearranged methods in order of expected behaviour, you have to fail
before you can succeed.