Refactor error handling for duplicate project name
Moved the error handling for creating new projects with already
existing names to the dialog from the presenter/fragment.

This allows us to use the standard error handling functionality from
the text field instead of using the AlertDialog.

The AlertDialog is somewhat cumbersome to work with and it requires
additional user interaction to dismiss, where the text field error
message will disappear when the user continues typing.