Notify user of dialog failure
Notify the user that the dialog for creating new projects could not be
displayed. The real reason for the failure is to technical to display
to the user, hence the unknown error message.

And, the listener should always be available in the production version,
i.e. this should just be seen as developer feedback.