Restructured the xml files
Added additional new lines between elements within the xml files to
improve readability, and added new line at the end of the file to the
files that were missing it.

The colors, dimens, and strings files are not separated between each
element, since it makes more sense to keep some elements grouped.