Added registered column to the time table
Added an additional column to the time table to enable support for
marking a row as registered. The commit also include the updated
database structure aswell as upgrade script for existing databases.

Since the SQLite library do not support dropping columns from a table
there is no downgrade available for this upgrade.