Display time marked as registered
Display the time marked as registered with a green background color.
The color seemed as a logical choice, green as in completed.

The current implementation relies on activating the view when the time
have been marked as registered.

This approach is not optimal since activating do not seem logical, but
the states available for the default view is somewhat limited.