Timesheet group fails to mark as registered
When the timesheet view have multiple expanded groups, and the user
marks the last unregistered item the group fails to be marked as
registered. If the group is any other than the first group.

The issue is related to the internal position system of the expandable
list adapter. I.e. the group position do not always match to the
correct adapter position, since its offset by child items.