Suppress warning of inverted boolean
The inspect code complains about 'Boolean method is always inverted'
for the methods `isViewAttached` and `has`.

The method names have been chosen because it better defines the flow
of the code than the negative counterpart, i.e. `isViewNotAttached`
and `hasNot`. Also, the shorter method name is preferred.

The warning have been suppressed with annotation.