Read/write hide preference for registered time
To enable the user to hide/show registered time, we have to store the
selected preference in a global container. This will also ease the
communication between the user interface and the data layer.

By using a simple wrapper around SharedPreferences for each of the
available preference we have a clean and simple interface to work with.