Refresh timesheet when hide preference is changed
To give the user direct feedback about the hide preference change, the
view have to refresh, i.e. reload the data.

When the 'Hide registered' menu item is checked/unchecked the view
`refresh`-method will clear the data and initiate a new load.

To avoid possible `NullPointerException`, the creation of the timesheet
fragment have been refactored into a getter method. The getter method
will create the fragment when needed.