Request read permission to external storage
To allow users to restore the backups we need to be able to check if
any backups are available.

By displaying a permission notification and allowing the user to
initiate the permission request, we can check available backups.

There is no need to do additional permission checks when running the
restore, since the action is disabled if permission have been denied.