Request write permission to external storage
To allow users to backup data, we need permission to write to the
external storage.

By displaying a permission notification and allowing the user to
initiate the permission request, we can backup data.

If permission have been denied, we should accept this and not display
the message again until the user attempts another backup.