simplify `bin/build`
Since `bin/build` is almost always used with the `--no-sign` option,
this should be the default behaviour.

And, since there is no actual need for the sign related code to be in
the `bin/build` we can migrate it to a separate script. This will also
greatly simplify the `bin/build` script.