move backup/restore classes to separate package
to better illustrate the separation of the classes/concepts.

the idea behind the separation is to keep the lower level modules within
the data-package, i.e. implementations talking to the filesystem.

these lower level modules implements interfaces from the domain-package
which is then used in with the domain logic, i.e. interactors.

i'm still a bit conflicted about the data/backup package but it's good
enough, atleast for now.