remove usage of observables from repository layer
the interfaces for the repository layer is located within the domain
package and should not be dependant on external libraries.

also, the simplification will make the layer easier to test since the
test do not need to handle observables.

the return type for the remove method have been changed to return void,
there's no reason to return anything since nothing can full be restored.