migrate create project to NewProjectPresenter
instead of sending the project to the ProjectsPresenter via a callback,
the project should be created via the NewProjectPresenter and then sent
to the list of projects.

this will considerably simplify the flow for creating projects since it
will only involve on `onNext`-event, and the code is close together.

also, it reduces the concerns of the ProjectsPresenter.