move project time method to TimeRepository
to ease the migration from ProjectProvider to use cases the method for
getting time registered to a project needs to be moved.

the method have been moved to the TimeRepository and the calling code
have been modified. this solution is only temporary and will be replaced
with a proper query method.