get project timesheet via use case
the ProjectProvider have been marked as deprecated and is going to be
replaced with separate use cases.

the `getTimesheet`-method in the TimesheetPresenter have been refactored
to use a simplified use case for getting project timesheet.

the remaining method in the ProjectProvider have been migrated to the
TimeRepository. since the ProjectProvider is no longer used, it and all
of the references can be removed.