correct time summary calculation
if a timesheet item includes items that trigger an additional minute
when calculated separately, the summary calculation might be incorrect.

the summary calculation do not account for additional minutes on
separate items as it only summarize the total milliseconds.

when calculating the time interval summary, each item have to be
calculated separately before added to the total.