improve Time construction with Builder
constructing the Time object is harder than it should be. the number of
arguments and their order (given that multiple arguments are of the same
type) decrease readability.

lack of support for default values forces each argument to be specified
for each new construction (having multiple constructors with different
default values would decrease readability).

the Time object should be constructed with a Builder pattern.