remove usage of ListFragment from ProjectsView
the `ListFragment` acts as a delegate between the presenter and adapter,
it reveals to much details of the fragments underlying data structure.

the coupling between the presenter and adapter have to be separate. the
details of the data structure have to be hidden, only the details that
are needed should be made visible.

since the presenter is using the `ListFragment` methods, these have to
be replaced with new methods.

the `ListFragment` inheritance have been removed, which allows for the
removal unnecessary boilerplate code.

the `ListFragment` have been marked as deprecated to avoid unnecessary
coupling in the future.