extract implementations for DateIntervalFormat
instead of `DateIntervalFormat` having separate internal methods for
different types of formatting, the implementations should be extracted
to separate classes and `DateIntervalFormat` converted into a common
interface.

this allows for a better separation between the implementation and will
ease addition of new implementations.

also, since the `DateIntervalFormat` only deals with the presentation,
the classes/interface should move to the presentation-package.