move view selection to the adapter
the fragment should not keep track of the internal state of the adapter,
i.e. the selected items, this behaviour should be within the adapter.

also, when selecting an item, scrolled the item off-screen and scrolled
back, the correct item might not be visually selected. the correct item
would still be registered or removed, but the list selection would be
displayed incorrectly.

by keeping an internal set of the selected item(s) and checking wether
an item is contained within the set, the selection is a lot easier.

also, this approach simplifies adding support for multi-select.