refine behaviour of long click
to allow for a more consistent selection behaviour, the long click have
to be refined. a long click should no longer reselect an item, instead
it should only be used to activate the selection (with an initial item).

a tap should be used to either reselect or deselect an item.