correct ui issue when removing multiple items
when the user removes multiple items the user-interface will fail to
properly update, if the items are selected in a random order.

to correct the issue the items have be sorted before removal, this will
require the entire list of selected items.

also, the items have to be sorted in the reverse order (to remove items
from the bottom up), since removing an item will cause the indexes for
the following items to change.