remove code for project description
it is not possible to add a project description from the user-interface,
the code only adds unnecessary complexity and nothing of real value.

due to the fact that SQLite is not able to remove columns, and the
feature will be added in the future the structure for the database
should remain intact.