remove code for project archiving
it is not possible to archive a project from the user-interface, the
code only adds unnecessary complexity and nothing of real value.

due to the fact that SQLite is not able to remove columns, and the
feature will be added in the future, the structure of the database
should remain intact.