replace ArrayList-inheritance with delegation
the ArrayList-inheritance expose to much of the internal data structure.
also, it forces the sub classes to do proper handling of serialization,
aswell as override the `equals` and `hashCode` methods.

the inheritance have been replaced with an implementation of the bare
minimum number of methods needed to provide the expected functionality.