add letter icon for timesheet days
to enable a consistent selection behaviour for days, an element that
can act as a selector is needed since the default behaviour for both
click and long click on days will either expand or collapse the
interval listing.

the selector is added in the form of a icon with the first letter of
the group title, i.e. first letter of the day.