use desired scheduler for refreshing projects
the default scheduler for the `interval`-operator is `computation` but
the subscription for refreshing active projects is using `newThread`.

this causes issues when testing due to the fact that two schedulers are
involved in the interval (not actively but initially).

instead of using the default scheduler and later replacing it with the
`newThread`, the desired scheduler should be used with the method-call.