add GetProjectTimeSince interactor
the `GetProjectTimeSince` interactor will get the registered time for a
project since a defined starting point, i.e. last day, week, month.

the interactor is needed to allow for different time interval summary.

due to the dynamic nature of the start point calculation, the test also
needs to dynamically calculate the starting point.