only post change event on actual change
there is no need to post the time summary starting point change event
when the new value is the same as the previous value, i.e. no actual
change have occurred.

since the settings are read from "global scope", i.e. no instance to
mock to simulate different values, and the default value is month the
test with month have to be replaced to simulate "no change"-event.