use explicit sorting for projects
when fetching the projects from the database no explicit sorting is
used (though it would seem that the projects are sorted by name).

this becomes an issue with multiple projects since the user-interface
update mechanism use a binary search based on the id.

if the projects are not sorted by id, the project might not be found
and the user-interface will fail to update the project.