add basic coverage unit tests for WorkerProvider
the main purpose of the tests is to get the basic behaviour under test
to ease incremental improvements in the future.

to enable testing of the interactions with the database the openHelper
field have to be exposed to the package. this will allow the tests to
swap the instance with a mock.