suppress notification errors from invalid project state
in, what seems like, random occurrences the pause and resume services
are called with a project that's in an invalid state for the called
service.

i.e. the pause service is called with an inactive project and the resume
service is called with an already active project.

from what i can tell, the user action of pausing or resuming a project
is performed and then the dismissed notification is displayed again,
with the now invalid project state.

since the internal behaviour is correct, in respect to the user action,
the state error can/should be suppressed pending more data.