depend on ClockIn use case in ClockActivityChange
if the user accidentally double tap the clock activity change toggle
(when the project is inactive), two new separate active time entries
are created.

this will cause the project to register double time. i.e. when pausing
and then resuming the project, the start time for the second interval
will be the same as the first interval.

this is caused due to the fact that the `ClockActivityChange` depend on
the project state from the presentation layer, and not the actual state
from the data store.

the `ClockActivityChange` should depend on the `ClockIn` use case, which
will enforce valid project state before clock in.