perform actions with view using strong reference
the pattern used to perform actions with the view is as follows, check
that the view is still attached and then perform the action.

since the view is stored in a `WeakReference`, and the attached check
and action are not performed against an explicit strong reference, the
actual view reference might have been released between the calls.

instead of checking whether the view is still attached using the weak
reference, a strong reference should be used.

and, to reduce the needed boilerplate and duplication, the action should
be passed as a lambda which only gets called if the view is attached.

one downside of using lambdas, at least in the current configuration, is
the number of additional methods generated.