throttle project action events
with multiple quick taps on either of the projects an error will shown
because the data layer can't handle the number of events.

the event stream have to be throttled to prevent flooding the data layer
with additional events, at least for a brief moment.

this can be done with the `throttleFirst` operator from RxJava.