check max/min date before use
the usage of min/max date in the `DateTimePickerFragment` is incorrect,
since it do not have proper `null` handling.

the arguments to the setter methods can not to be `null`. however, not
calling the methods will cause the fields to use their default values,
which in this case is `null`.