suppress notification errors from invalid project state
in, what seems like, random occurrences the clock out service is called
with a project that's in an invalid state for the service.

attempts to reproduce this issue, on either of the available development
devices, have failed. however, it do occur a few times each week on
production devices.

both the pause and resume service displayed the same issue, although
more frequent, and have previously been suppressed in commit 507749a.