keep apk artifact between ci stages
the built apk artifact should be passed between each stage, i.e. build
stage produce the apk, test stage run tests against the built apk, and
deployment push the built and tested apk.

this will speed up the ci pipeline and ensure that the deployed artifact
have in fact been tested.

since the sonar analyze job is depending on testing and is running for
all tags, the test job is not needed for tags.