inject database into provider using Dagger
to enable testing with another database, e.g. in-memory database, the
instance have to be injected with Dagger.

due to an implementation detail of the start up process this injection
have to be done "on demand", the application have not been created when
the provider is initialized.

and, since the provider is called from different threads the calls have
to be syncronized.