handle backpressure for time sheet observables
when the observables emits multiple of items the subscription on the
main thread can't keep up and chokes causing a missing backpressure
exception, which in turn cause the observable to be terminated.

attempting to apply the `buffer` operator between `subscribeOn` and
`observeOn` seem to cause issues with the user interface (unsure what
the reason behind this is).

however, applying the `onBackpressureBuffer` operator seem to work.

only observables which are able to emit multiple items are affected by
this issue with backpressure.