sort time sheet results before passing them to ui
when registered items are hidden from the time sheet and the user is
attempting to register multiple additional items, not all of the items
are removed from the user-interface.

this is because the items are not sorted before they're removed, i.e.
the indexes for the items do not match the correct item after the first
one have been removed.