remove android integration test
the android integration test block remove usage of dagger, due to how
we initialize a clean database, i.e. inject dependencies on demand
within the getter.

i don't seem to be able to get this to work using koin since everything
i try will either not allow me to setup another database in test or
fails due to the fact that the providers `onCreate` method is called
before the applications.