reload notifications when application is resumed
if the user has an active project but the ongoing notification channel
is disabled. when the user enables the notification channel, the
ongoing notification will not be displayed until the next notification
is sent.

we should attempt to mimic the behaviour for when the application first
starts with an active project, i.e. the pause notification is sent. this
can be done by reloading the notifications in `HomeActivity.onResume`.
the performance (CPU/battery usage) is essentially non-existing for when
a single project is active.