remove dependency for robolectric
after updating the android gradle build told plugin to 3.4.0, the
robolectric dependency is causing issues. it seems to be an known issue
which might be solved by upgrading the robolectric dependency.

since the robolectric usage is not extended within the app, we should
opt for removing it and later add test for the missing parts.