reload notifications when activity is created
attempting to reload the notifications each time the activity is resumed
caused the application to crash after the device had been in sleep for a
while.

since the reason for the notification reload is to ensure notifications
for active projects are active after e.g. device reboot, it should be
sufficient to perform this action when the activity is created.

closes #179