allow time report day/item to recalculate summary
the first time the summary has been calculated, it's stored in the
object causing the same value to be used for later. when we attempt to
refresh the time summary the old value is still presented instead of
the new value.

in order to resolve this we should allow days and items to recalculate
their time summary values.