remove use of configure android environment
there's no need to use `configure-android-environment` since it no
longer installs the android api, and if the build tools is missing it
will be installed during compilation.

this will remove the need to keep two locations in sync regarding the
build tools version, which is more prone to change than the android api.