update references for target sdk version
in order to update the target sdk version we need to update the version
references in `app/build.gradle` and the docker image used for ci/cd.

update to version 29 also required two services to allow for nullable
intent in `onHandleIntent(Intent?)`.
update to version 29 also required two services to allow for