run refresh active time intervals on main thread
running the timer closure using a coroutine might be causing some issues
when the application is resumed from the background.

however, this is just speculations since i've yet to reproduce it in an
actual test with the debugger attached, but the call stack seems to
indicate that it's related to how we call the refresh from a coroutine.

since the refresh calculations are not too complicated it should be fine
running these on the main thread without any performance impact.

and, hopefully the issue is either gone or we get another easier to
follow stack trace.