use time report day for checking expand/collapse state
in order to use `TimeReportWeek` as the base we need to check whether a
day is expanded or collapsed based on the day and not the position (as
the position will reference the week and not the day).

this also requires us to notify that the entire data set has changed when
a day has been expanded or collapsed, and not just the individual day.