include adaptive launcher icon
the launcher icon on the pixel launcher is incorret, it uses a scaled
down version of the icon within a circle. to resolve this issue we need
to use an adaptive icon and declare the `roundIcon`.

since our min api level is 27 and the adaptive icon, by default, create
an directory for v26 and above we need to add a lint baseline to ignore
the lint warning. merging the directory to a `minmap-anydpi` folder did
not seem to work on the Nexus 6P, i.e. api 28.

closes #229