observe using viewLifecyclerOwner instead of LifecyclerOwner
using `LivecycleOwner` to observe values might leak the reference, i.e.
fragment or activity. instead we should use the `viewLifecycleOwner` of
the fragment when observing values.

using the `viewLifecycleOwner` requires that the view have been created,
which might not be the case for some places where we start observing. in
order to ensure consistency we should always observe the view model from
fragments in `onActivityCreated`, which is also referenced in the
documentation.

closes #232