check if project is available on background thread
instead of running the "if project is available"-check on the main
thread we should run it on a background thread.

this change causes us to update the coroutine version used, in order
to access the `TestCoroutineDispatcher`, as well as lifecycle and
fragment for the coroutine enabled live data construct.

closes #145