force loading of projects on background thread
in order to ensure that the application is responsive we need to ensure
that all database operations are performed on a background thread.

this is done by converting the related use cases and repository methods
to suspending methods and then using the view model scope when loading
data via the data source.

however, this introduces an issue where the data observation from the
fragment is not able detect whether the list is empty or not (because
the list is empty when we submit it, but it later gets populated). this
causes an issue with displaying the empty view image which i've
temporarily disabled, this has been registered as a separate issue to be
resolved at a later stage (see #239).