improve calculation for "set to start of week"
during the weeks surrounding the new year (2019-2020) an issue presented
itself with the time report grouping (it was probably an issue all along
but never fully manifested itself).

the `setToStartOfWeek` did not fully handle the sunday vs. monday as the
first day of the week resulting in incorrect grouping with fixed values.