resolve issue with failing tests for time report
when running the instrumented tests certain parts of the time report
data layer will fail due to an issue with time zones. when attempting
to get time intervals within the same week there is an offset causing
the time report to be fetch on different weeks.

using both `localtime` in the sql layer and removing the explicit use
of time zones in the "set to point in time" methods, we should be able
to always fetch everything using localtime.

closes #244