improve support for clock out over multiple days
instead of creating a single time interval that can span multiple days,
we should split the time intervals into days.

along with this change we need to adjust the "end of week" property that
is used in several tests as the new behaviour would create additional
time intervals as it might span multiple days.