run unit and instrumented tests using debug build
using `testCoverageEnabled true` for the release build will cause the
build to be debuggable, which might prevent it from being uploaded to
google play.

the configuration is required to get the test coverage when running
instrumented tests. in order to handle both of these scenarios we need
to build both a debug and a release build, test the debug build and
deploy the release build.

however, we still need to use the instrumented tests to ensure that the
app will start and not crash due to missing classes after minification,
i.e. both debug and release need the same configuration in ci.