use instrumented build type when running tests
using the debug build type did not produce the expected behaviour when
running the instrumented tests in firebase test labs.

the expected behaviour was that the test would run as with the release
build, but the emulator crashed due to a `NoSuchMethodError` exception.

the exception message reference the flipper dependency, which is only
present in the debug build type.

using a separate build type for the tests based on the release, should
resolve the issue.