expose generated coverage reports as artifacts
there seems to be some issues with the coverage reports when pushing
sonarcloud, the instrumented tests are not included.

we need to expose the coverage reports sent to sonarcloud as artifacts
in order to inspect them and verify if the issue is with the report or
with the plugin in sonarcloud.

when the reports were sent to an on-prem installation of sonarqube, the
coverage report was a lot higher than reported from sonarcloud.

in order to ease inspection, the html report have also been enabled.