update android build tools to version 4.0.0
along with the android build tools to version 4.0.0, we also need to
update gradle to version 6.1.1 as well as migrate crashlytics to use
firebase instead of fabric.

the reason for the crashlytics to firebase migration is due to an issue
when attempting to compile the core and core-test module in a separate
gradle container. when running `gradle core:assemble core-test:assemble`
in a `gradle:6.1.1-jdk8` docker container along with android build tools
version 4.0.0, the command will also run the `app:assembleRelease` task
which will fail due to the android sdk not being available.