Revert "use jdk11 when sending sonar analysis"
attempting to use the jdk11 image for sonar analysis will fail because
the android environment is not installed. and, using the openjdk11 as
base for the android environment image will fail because incompatible
changes was made in jdk9, e.g. installation of tools via sdkmanager will
fail because `javax/xml/bind/annotation/XmlSchema` is not found.

This reverts commit 217788dfe551790b1086ae0a2f9a81a67f73cdc8.