Post analysis to sonarcloud using sonar-scanner-cli
Tnstead of using the sonarqube gradle plugin we need to migrate to using
the `sonar-scanner-cli`. The reason for this is that support for JDK 1.8
have been deprecated and we are unable to upgrade the container used for
the android-environment to anything above JDK 1.8 since the sdkmanager
will crash because `javax/xml/bind/annotation/XmlSchema` is missing.

In order to migrate to using sonar-scanner-cli, we first need to convert
the existing configuration to the `sonar-project.properties`. There are
a few properties that can not be included in the properties file due to
both security concerns and the dynamic nature of the value, i.e. the
authentication information and project version needs to be passed as
additional arguments.

The project version also needs to be passed from the `core:build` job
as an environment variable. This is because I was unable to extract the
version from the `showVersion` output using the tools available in the
alpine variant of the container. The regexp did not work with `grep`
and `awk` do not support capture groups, anything more elaborate would
be more difficult to maintain than passing an environment variable.

Closes #63 and #254