update android build tools to 7.0.0-alpha08
the versioning scheme for the android build tools gradle plugin is
going to change to be more similar to the gradle version, i.e. the
stable version of the gradle plugin will most likely require the
version 7.0 of gradle.

as for the alpha release, it requires the use of gradle 6.8.2, which
in turn requires jdk version 11.