remove use of instrumented build type
the use of the `instrumented` build type was intended to be a
combination of both the `debug` and `release` build types, i.e.
include proguard configuration and enable test coverage.

this can be resolved with conditional configuration for the
debug build type, i.e. enable proguard in debug if `CI` is present.